requirements.txt
scikit-learn
tensorflow-gpu
matplotlib
tqdm